refactor(fpss): extract decode.rs
Move decode_frame (the per-frame dispatch from the I/O loop) and the
EMPTY_SYMBOL sentinel out of fpss/mod.rs into a dedicated decode submodule.
The 8-field and 16-field trade-mapping tests (and their FIT encoding
helpers) move alongside. No functional change -- the only visibility bump
was decode_frame itself becoming pub(super) so io_loop can still call it.
